refactor: move Asteroids to root
