+++ app/asteroids/page.tsx
@@ -1,12 +1,100 @@
-import Basket from '../components/Basket';
-import Main from '../components/Main';
-import AsteroidListServer from '../components/AsteroidListServer';
+'use client';
+import { useEffect, useRef, useState } from 'react';
+import AsteroidList from '../components/AsteroidList';
 
-export default async function Asteroids() {
+export default function Asteroids() {
+  const [asteroids, setAsteroids] = useState<AsteroidOnClient[]>([]);
+  const [loading, setLoading] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
+  const observerTargetDown = useRef(null);
+  const observerTargetUp = useRef(null);
+
+  useEffect(() => {
+    const observerDownUnobserve = () => {
+      if (observerTargetDown.current) {
+        observerDown.unobserve(observerTargetDown.current);
+      }
+    };
+    const observerDownObserve = () => {
+      if (observerTargetDown.current) {
+        observerDown.observe(observerTargetDown.current);
+      }
+    };
+    const observerUpUnobserve = () => {
+      if (observerTargetUp.current) {
+        observerUp.unobserve(observerTargetUp.current);
+      }
+    };
+    const observerUpObserve = () => {
+      if (observerTargetUp.current) {
+        observerUp.observe(observerTargetUp.current);
+      }
+    };
+    async function getData(move = '') {
+      setLoading(true);
+      try {
+        const response = !move ? await fetch('/api/asteroids') : await fetch(`/api/asteroids?move=${move}`);
+        if (!response.ok) {
+          throw new Error('Failed to fetch data');
+        }
+        const data: { asteroidList: AsteroidOnClient[], isStart: boolean; } = await response.json();
+        const { asteroidList, isStart } = data;
+        setAsteroids(asteroidList);
+        setTimeout(() => {
+          if (!isStart) {
+            observerUpObserve();
+          }
+          observerDownObserve();
+        }, 900);
+      } catch (error: Error | unknown) {
+        let message = 'Failed to fetch data';
+        if (error instanceof Error) message = error.message;
+        setErrorMessage(message);
+      } finally {
+        setLoading(false);
+      }
+    }
+    const options = {
+      threshold: 1,
+      root: document,
+      rootMargin: "20px",
+    };
+    const observerDown = new IntersectionObserver(
+      entries => {
+        if (entries[0].isIntersecting && !errorMessage) {
+          observerDownUnobserve();
+          getData('down');
+        }
+      }, options
+    );
+    const observerUp = new IntersectionObserver(
+      entries => {
+        if (entries[0].isIntersecting && !errorMessage) {
+          observerUpUnobserve();
+          window.scrollTo({
+            top: 50,
+            left: 0,
+            behavior: "smooth",
+          });
+          getData('up');
+        }
+      }, options
+    );
+    async function firstLoading() {
+      await getData('down');
+    }
+    if (!asteroids.length) firstLoading();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [observerTargetDown]);
   return (
     <div>
-      <Main />
-      {/* <AsteroidListServer /> */}
+      <div ref={observerTargetUp}></div>
+      {errorMessage && <div>{errorMessage}</div>}
+      <h1>Asteroids</h1>
+      {asteroids && asteroids.length > 0 && <AsteroidList asteroids={asteroids} />}
+      {loading && <div>Loading...</div>}
+      {errorMessage && <div>{errorMessage}</div>}
+      <div ref={observerTargetDown}></div>
     </div>
   );
 }
