--- app/components/Basket.tsx
@@ -1,5 +1,0 @@
-export default async function Basket() {
-  return (
-    <div>Basket</div>
-  );
-}
