+++ app/components/Cart.tsx
@@ -1,0 +1,5 @@
+export default async function Cart() {
+  return (
+    <div>Cart</div>
+  );
+}
